change  fp32 to fp16 for weight (#244)

+++ paddle/fluid/distributed/ps/table/common_graph_table.cc
@@ -194,7 +194,7 @@ int32_t GraphTable::Load_to_ssd(const std::string &path,
   std::vector<uint64_t> edge_array[task_pool_size_];  // edge id list
 
   // get edge weight
-  std::vector<float> weight_array[task_pool_size_];  // neighbor weight list
+  std::vector<half> weight_array[task_pool_size_];  // neighbor weight list
 
   for (size_t i = 0; i < bags.size(); i++) {
     if (bags[i].size() > 0) {
