+++ paddle/fluid/distributed/ps/table/graph/graph_edge.cc
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #include "paddle/fluid/distributed/ps/table/graph/graph_edge.h"
-
 #include <cstring>
 namespace paddle {
 namespace distributed {
@@ -24,7 +23,7 @@ void GraphEdgeBlob::add_edge(int64_t id, float weight = 1) {
 
 void WeightedGraphEdgeBlob::add_edge(int64_t id, float weight = 1) {
   id_arr.push_back(id);
-  weight_arr.push_back(weight);
+  weight_arr.push_back((half)weight);
 }
 }  // namespace distributed
 }  // namespace paddle
